+++ TLSharp.Core/Network/Requests/AckRequest.cs
@@ -2,20 +2,27 @@
 using System.Collections.Generic;
 using System.IO;
 using TeleSharp.TL;
-using TLSharp.Core.Utils;
 
-namespace TLSharp.Core.Requests
+namespace TLSharp.Core.Network.Requests
 {
-    public class PingRequest : TeleSharp.TL.TLMethod
+    public class AckRequest : TLMethod
     {
-        public PingRequest()
+        private readonly List<ulong> _msgs;
+
+        public AckRequest(List<ulong> msgs)
         {
+            _msgs = msgs;
         }
 
         public override void SerializeBody(BinaryWriter writer)
         {
-            writer.Write(Constructor);
-            writer.Write(Helpers.GenerateRandomLong());
+            writer.Write(0x62d6b459); // msgs_ack
+            writer.Write(0x1cb5c415); // Vector
+            writer.Write(_msgs.Count);
+            foreach (ulong messageId in _msgs)
+            {
+                writer.Write(messageId);
+            }
         }
 
         public override void DeserializeBody(BinaryReader reader)
@@ -28,11 +35,15 @@ namespace TLSharp.Core.Requests
             throw new NotImplementedException();
         }
 
+        public override bool Confirmed => false;
+
+        public override bool Responded { get; }
+
         public override int Constructor
         {
             get
             {
-                return 0x7abe77ec;
+                return 0x62d6b459;
             }
         }
     }
