+++ TLSharp.Core/Network/Requests/PingRequest.cs
@@ -1,0 +1,38 @@
+﻿using System;
+using System.IO;
+using TeleSharp.TL;
+using TLSharp.Core.Utils;
+
+namespace TLSharp.Core.Network.Requests
+{
+    public class PingRequest : TLMethod
+    {
+        public PingRequest()
+        {
+        }
+
+        public override void SerializeBody(BinaryWriter writer)
+        {
+            writer.Write(Constructor);
+            writer.Write(Helpers.GenerateRandomLong());
+        }
+
+        public override void DeserializeBody(BinaryReader reader)
+        {
+            throw new NotImplementedException();
+        }
+
+        public override void DeserializeResponse(BinaryReader stream)
+        {
+            throw new NotImplementedException();
+        }
+
+        public override int Constructor
+        {
+            get
+            {
+                return 0x7abe77ec;
+            }
+        }
+    }
+}
