+++ TLSharp.Core/TLSharp.Core.csproj
@@ -71,8 +71,8 @@
     <Compile Include="Network\TcpTransport.cs" />
     <Compile Include="Network\Exceptions\UserMigrationException.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="Requests\AckRequest.cs" />
-    <Compile Include="Requests\PingRequest.cs" />
+    <Compile Include="Network\Requests\AckRequest.cs" />
+    <Compile Include="Network\Requests\PingRequest.cs" />
     <Compile Include="Utils\UploadHelper.cs" />
     <Compile Include="Session.cs" />
     <Compile Include="TelegramClient.cs" />
@@ -88,6 +88,7 @@
       <Name>TeleSharp.TL</Name>
     </ProjectReference>
   </ItemGroup>
+  <ItemGroup />
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it. 
        Other similar extension points exist, see Microsoft.Common.targets.
